requirements.txt
asgiref==3.7.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cors==1.0.1
dj-database-url==2.0.0
Django==3.2.7
django-cors-headers==3.9.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.12.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
flask-jwtlogin==0.2.1
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==10.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
waitress==2.1.2
Werkzeug==2.2.3
whitenoise==6.5.0
zope.event==4.6
zope.interface==6.0
